refactor: break down UnifiedCard into modular components
- Refactor monolithic 897-line UnifiedCard into 7 focused components
- Extract sections: ThumbnailSection, HeaderSection, TechnologiesSection,
  AIToolsSection, AchievementsSection, TechnicalDetailsButton, FooterSection
- Create comprehensive TypeScript types with proper interfaces
- Implement utility functions for styling and helper logic
- Maintain all existing functionality while improving maintainability
- Achieve successful build compilation with zero errors

Components created:
- ThumbnailSection: Image handling, fallbacks, and status badge overlay
- HeaderSection: Title, subtitle, description with work/project layouts
- TechnologiesSection: Technology badges with limiting functionality
- AIToolsSection: AI tool badges with limiting functionality
- AchievementsSection: Professional achievements and highlights display
- TechnicalDetailsButton: Prominent CTA button for technical details
- FooterSection: Metadata, timeline, and action buttons
- Utils: Card styling, size configuration, and helper functions
- Types: Comprehensive TypeScript interfaces for all component props

This refactoring improves code organization, reusability, and maintainability
while preserving the exact same user experience and functionality.
